add user permissions to blog posts

+++ cmsapp/views.py
@@ -4,6 +4,8 @@ from .models import Post
 from .forms import PostForm
 from django.utils.text import slugify
 from django.contrib.auth.decorators import login_required
+from profile.models import UserProfile
+
 
 # Create your views here.
 def index(request):
@@ -28,11 +30,14 @@ def blogpost(request, slug):
 @login_required
 def create_post(request):
     form = PostForm
+    user_id = request.user
+    user_profile = UserProfile.objects.get(user_id=user_id)
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES)
         if form.is_valid():
             post = form.save(commit=False)
             post.slug = slugify(post.title)
+            post.author = user_profile
             post.save()
             messages.info(request, 'Article was created successfully')
             return redirect('index')
@@ -42,28 +47,38 @@ def create_post(request):
 
 @login_required
 def update_post(request, slug):
+    user_id = request.user
+    user_profile = UserProfile.objects.get(user_id=user_id)
     post = Post.objects.get(slug=slug)
-    form = PostForm(instance=post)
-    if request.method == 'POST':
-        form = PostForm(request.POST, request.FILES, instance=post)
-        if form.is_valid():
-            form.save()
-            messages.info(request, 'Article was updated successfully')
-            return redirect('blogpost', slug=post.slug)
-    context = {'form': form}
-    return render(request, 'create_post.html', context)
+    if post.author == user_profile:
+        form = PostForm(instance=post)
+        if request.method == 'POST':
+            form = PostForm(request.POST, request.FILES, instance=post)
+            if form.is_valid():
+                form.save()
+                messages.info(request, 'Article was updated successfully')
+                return redirect('blogpost', slug=post.slug)
+        context = {'form': form}
+        return render(request, 'create_post.html', context)
+    else:
+        return redirect('index')
 
 
 @login_required
 def delete_post(request, slug):
+    user_id = request.user
+    user_profile = UserProfile.objects.get(user_id=user_id)
     post = Post.objects.get(slug=slug)
-    form = PostForm(instance=post)
-    if request.method == 'POST':
-        post.delete()
-        messages.info(request, 'Article was deleted successfully')
+    if post.author == user_profile:
+        form = PostForm(instance=post)
+        if request.method == 'POST':
+            post.delete()
+            messages.info(request, 'Article was deleted successfully')
+            return redirect('index')
+        context = {'form': form}
+        return render(request, 'delete_post.html', context)
+    else:
         return redirect('index')
-    context = {'form': form}
-    return render(request, 'delete_post.html', context)
 
 
 def about(request):
